requirements.txt
streamlit
pandas
openpyxl
psycopg2-binary
numpy
streamlit_option_menu